feat/user-edit
User's can now edit their accounts.

I likely would need to make changes to the API to determine if the
request is for the user's own account, if true then return additional
fields such as Email.

This would allow admins + users to edit their own email. For now I'll
just be ignoring it.

Closes #15